+++ eks_fargate/CHANGELOG.md
@@ -1,1 +1,5 @@
 # CHANGELOG - eks_fargate
+## 1.0.0 / 2020-02-22
+
+* [Added] Introducing the eks_fargate check. See [#5417](https://github.com/DataDog/integrations-core/pull/5417).
+